fix: only trigger if member.roles is present

+++ src/bot/listeners/client/guildMemberUpdate.js
@@ -12,11 +12,13 @@ class GuildMemberUpdateListener extends Listener {
 	async exec(oldember, newMember) {
 		const roleState = this.client.settings.get(newMember.guild, 'roleState');
 		if (roleState) {
-			const roles = newMember.roles.filter(role => role.id !== newMember.guild.id).map(role => role.id);
-			if (roles.length) {
-				await this.client.db.models.role_states.upsert({ guild: newMember.guild.id, user: newMember.id, roles });
-			} else {
-				await this.client.db.models.role_states.destroy({ where: { guild: newMember.guild.id, user: newMember.id } });
+			if (newMember.roles) {
+				const roles = newMember.roles.filter(role => role.id !== newMember.guild.id).map(role => role.id);
+				if (roles.length) {
+					await this.client.db.models.role_states.upsert({ guild: newMember.guild.id, user: newMember.id, roles });
+				} else {
+					await this.client.db.models.role_states.destroy({ where: { guild: newMember.guild.id, user: newMember.id } });
+				}
 			}
 		}
 	}
